HW7/DrawingModel/DrawingModel/CommandManager.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace DrawingModel
{
    public class CommandManager
    {
        Stack<ICommand> _undo = new Stack<ICommand>();
        Stack<ICommand> _redo = new Stack<ICommand>();

        //Execute
        public void Execute(ICommand command)
        {
            command.Execute();
            _undo.Push(command);
            _redo.Clear();
        }

        //Undo
        public void Undo()
        {
            if (_undo.Count <= 0)
                throw new Exception("");
            ICommand command = _undo.Pop();
            _redo.Push(command);
            command.CancelExecute();
        }

        //Redo
        public void Redo()
        {
            if (_redo.Count <= 0)
                throw new Exception("");
            ICommand command = _redo.Pop();
            _undo.Push(command);
            command.Execute();
        }

        public bool IsRedoEnabled
        {
            get
            {
                return _redo.Count != 0;
            }
        }

        public bool IsUndoEnabled
        {
            get
            {
                return _undo.Count != 0;
            }
        }
    }
}
